+++ src/core/unif.ml
@@ -8,13 +8,21 @@ open Common
 open Error
 open Term
 open LibTerm
-open Env
 open Print
 open Debug
 
 (** Logging function for unification. *)
-let log_unif = new_logger 'u' "unif" "unification"
-let log_unif = log_unif.logger
+let logger_unif = new_logger 'u' "unif" "unification"
+let log_unif = logger_unif.logger
+
+(** [type_app a ts] returns [Some(u)] where [u] is a type of [add_args x ts]
+   where [x] is any term of type [a] if [x] can be applied to at least
+   [List.length ts] arguments, and [None] otherwise. *)
+let rec type_app : ctxt -> term -> term list -> term option = fun ctx a ts ->
+  match Eval.whnf ctx a, ts with
+  | Prod(_,b), t::ts -> type_app ctx (Bindlib.subst b t) ts
+  | _, [] -> Some a
+  | _, _ -> None
 
 (** Exception raised when a constraint is not solvable. *)
 exception Unsolvable
@@ -23,7 +31,7 @@ exception Unsolvable
    using declared unification rules. *)
 let try_unif_rules : ctxt -> term -> term -> constr list option =
   fun ctx s t ->
-  if !log_enabled then log_unif "check unif_rules for %a" pp_constr (ctx,s,t);
+  if !log_enabled then log_unif "check unif_rules";
   let exception No_match in
   let open Unif_rule in
   try
@@ -37,130 +45,86 @@ let try_unif_rules : ctxt -> term -> term -> constr list option =
       | Some(_)    -> assert false (* Everything should be matched *)
       | None       -> raise No_match
     in
-    let subpbs = List.map (fun (t,u) -> (ctx,t,u)) (unpack rhs) in
-    let pp_subpbs = List.pp pp_constr "; " in
-    if !log_enabled then log_unif (gre "get [%a]") pp_subpbs subpbs;
-    Some(subpbs)
+    let cs = List.map (fun (t,u) -> (ctx,t,u)) (unpack rhs) in
+    if !log_enabled then log_unif "rewrites to:%a" pp_constrs cs;
+    Some cs
   with No_match ->
     if !log_enabled then log_unif "found no unif_rule";
     None
 
-(** [nl_distinct_vars ctx ts] checks that [ts] is made of variables  [vs] only
-    and returns some copy of [vs] where variables occurring more than once are
-    replaced by fresh variables.  Variables defined in  [ctx] are unfolded. It
-    returns [None] otherwise. *)
-let nl_distinct_vars
-    : ctxt -> term array -> (tvar array * tvar StrMap.t) option =
-  fun ctx ts ->
-  let exception Not_a_var in
-  let open Stdlib in
-  let vars = ref VarSet.empty
-  and nl_vars = ref VarSet.empty
-  and patt_vars = ref StrMap.empty in
-  let rec to_var t =
-    match Ctxt.unfold ctx t with
-    | Vari(v) ->
-        if VarSet.mem v !vars then nl_vars := VarSet.add v !nl_vars
-        else vars := VarSet.add v !vars;
-        v
-    | Symb(f) when f.sym_name <> "" && f.sym_name.[0] = '$' ->
-        (* Symbols replacing pattern variables are considered as variables. *)
-        let v =
-          try StrMap.find f.sym_name !patt_vars
-          with Not_found ->
-            let v = Bindlib.new_var mkfree f.sym_name in
-            patt_vars := StrMap.add f.sym_name v !patt_vars;
-            v
-        in to_var (Vari v)
-    | _ -> raise Not_a_var
-  in
-  let replace_nl_var v =
-    if VarSet.mem v !nl_vars
-    then Bindlib.new_var mkfree (Bindlib.name_of v)
-    else v
-  in
-  try
-    let vs = Array.map to_var ts in
-    let vs = Array.map replace_nl_var vs in
-    (* We remove non-linear variables from [!patt_vars] as well. *)
-    let fn n v m = if VarSet.mem v !nl_vars then m else StrMap.add n v m in
-    let map = StrMap.fold fn !patt_vars StrMap.empty in
-    Some (vs, map)
-  with Not_a_var -> None
+(** [instantiable ctx m ts u] tells whether, in a problem [m[ts]=u], [m] can
+   be instantiated. It does not check whether the instantiation is closed
+   though. *)
+let instantiable : ctxt -> meta -> term array -> term -> bool =
+  fun ctx m ts u -> not (Meta.occurs m u) && nl_distinct_vars ctx ts <> None
 
-(** [sym_to_var m t] replaces in [t] every symbol [f] by a variable according
-   to [m]. *)
-let sym_to_var : tvar StrMap.t -> term -> term = fun m ->
-  let rec to_var t =
-    match unfold t with
-    | Symb(f)     -> (try Vari (StrMap.find f.sym_name m) with Not_found -> t)
-    | Prod(a,b)   -> Prod (to_var a, to_var_binder b)
-    | Abst(a,b)   -> Abst (to_var a, to_var_binder b)
-    | LLet(a,t,u) -> LLet (to_var a, to_var t, to_var_binder u)
-    | Appl(a,b)   -> Appl(to_var a, to_var b)
-    | Meta(m,ts)  -> Meta(m, Array.map to_var ts)
-    | Patt(_)     -> assert false
-    | TEnv(_)     -> assert false
-    | TRef(_)     -> assert false
-    | _           -> t
-  and to_var_binder b =
-    let (x,b) = Bindlib.unbind b in
-    Bindlib.unbox (Bindlib.bind_var x (lift (to_var b)))
-  in to_var
-
-(** [instantiation ctx m ts u] checks whether, in a problem [m[ts]=u], [m] can
-    be instantiated and returns the corresponding instantiation. It does not
-    check whether the instantiation is closed though. *)
+(** [instantiation ctx m ts u] tells whether, in a problem [m[ts]=u], [m] can
+   be instantiated and returns the corresponding instantiation, simplified. It
+   does not check whether the instantiation is closed though. *)
 let instantiation : ctxt -> meta -> term array -> term ->
   tmbinder Bindlib.box option = fun ctx m ts u ->
   if not (Meta.occurs m u) then
     match nl_distinct_vars ctx ts with
-    | None       -> None
-    | Some(vs,m) ->
-        let u = if StrMap.is_empty m then u else sym_to_var m u in
+    | None -> None
+    | Some(vs, map) ->
+        let u = Eval.simplify (sym_to_var map u) in
         Some (Bindlib.bind_mvar vs (lift u))
   else None
 
 (** Checking type or not during meta instanciation. *)
 let do_type_check = Stdlib.ref true
 
-(** [instantiate ctx m ts u] check whether, in a problem [m[ts]=u], [m] can be
-    instantiated and, if so, instantiate it. *)
-let instantiate : ctxt -> meta -> term array -> term -> constr list -> bool =
-  fun ctx m ts u initial ->
-  if !log_enabled then
-    log_unif "try instantiate %a" pp_constr (ctx,Meta(m,ts),u);
+(** Initial set of constraints. *)
+let initial : constr list Stdlib.ref = Stdlib.ref []
+
+(** [is_initial c] checks whether [c] occurs in [!initial]. *)
+let is_initial c = List.exists (LibTerm.eq_constr c) Stdlib.(!initial)
+
+(** [instantiate ctx m ts u] check whether, in a problem [m[ts] ≡ u], [m] can
+   be instantiated and, if so, instantiate it. *)
+let instantiate : ctxt -> meta -> term array -> term -> bool =
+  fun ctx m ts u ->
+  if !log_enabled then log_unif "try instantiate";
   match instantiation ctx m ts u with
   | Some(bu) when Bindlib.is_closed bu ->
       let do_instantiate() =
-        if !log_enabled then log_unif (yel "%a ≔ %a") pp_meta m pp_term u;
+        if !log_enabled then log_unif (red "%a ≔ %a") pp_meta m pp_term u;
         Meta.set m (Bindlib.unbox bu); true
       in
       if Stdlib.(!do_type_check) then
         begin
           let typ_mts =
-            match Infer.type_app ctx !(m.meta_type) (Array.to_list ts) with
+            match type_app ctx !(m.meta_type) (Array.to_list ts) with
             | Some a -> a
             | None -> assert false
           in
+          if !log_enabled then log_unif "check typing";
           match Infer.check_noexn [] ctx u typ_mts with
           | None ->
               if !log_enabled then log_unif "typing condition failed"; false
           | Some [] -> do_instantiate()
           | Some cs ->
-              let is_initial c = List.exists (LibTerm.eq_constr c) initial in
               let cs = List.filter (fun c -> not (is_initial c)) cs in
               if cs = [] then do_instantiate()
               else
                 (if !log_enabled then
-                   (let constr ppf = Format.fprintf ppf "\n; %a" pp_constr in
-                    log_unif "typing generated new constraints:%a"
-                      (List.pp constr "") cs);
+                   log_unif
+                     "failed because typing generated new constraints:%a"
+                     pp_constrs cs;
                  false)
         end
       else do_instantiate()
-  | _ ->
-      if !log_enabled then log_unif "variable condition failed"; false
+  | i ->
+      if !log_enabled then
+        begin
+          match i with
+          | None ->
+              if not (Meta.occurs m u) then log_unif "occur check failed"
+              else log_unif "arguments are not distinct variables: %a"
+                     (Array.pp pp_term " ") ts
+          | Some _ -> log_unif "not closed"
+        end;
+      false
 
 (** [error t1 t2]
 @raise Unsolvable. *)
@@ -168,35 +132,67 @@ let error : term -> term -> 'a = fun t1 t2 ->
   fatal_msg "\n[%a]\nand\n[%a]\nare not convertible.\n" pp_term t1 pp_term t2;
   raise Unsolvable
 
+(** [add_constr c cs] adds [c] to [!initial] and returns [c::cs]. *)
+let add_constr : constr -> constr list -> constr list = fun c cs ->
+  if !log_enabled then log_unif (mag "%a") pp_constr c;
+  Stdlib.(initial := c::!initial);
+  c::cs
+
+let add_constr_fold cs c = add_constr c cs
+
+(** [add_unif_rule_constr cs (ctx,t,u)] adds to [cs] the constraint
+   [(ctx,t,u)] as well as the constraint [(ctx,a,b)] where [a] is the type of
+   [t] and [b] the type of [u] if they can be infered. *)
+let add_unif_rule_constr : constr list -> constr -> constr list =
+  fun cs ((ctx,t,u) as c) ->
+  match Infer.infer_noexn [] ctx t with
+  | None ->
+      begin
+        match Infer.infer_noexn [] ctx u with
+        | None -> add_constr c cs
+        | Some (_,csu) ->
+            List.fold_left add_constr_fold (add_constr c cs) csu
+      end
+  | Some (a,cst) ->
+      match Infer.infer_noexn cst ctx u with
+      | None -> List.fold_left add_constr_fold (add_constr c cs) cst
+      | Some (b,cstu) ->
+          let cs = List.fold_left add_constr_fold (add_constr c cs) cstu in
+          if Eval.eq_modulo ctx a b then cs else add_constr (ctx,a,b) cs
+
 (** [add_to_unsolved t1 t2 p] checks whether [t1] is equivalent to [t2]. If
-   not, then it tries to apply unification rules on the problem [t1 = t2]. If
+   not, then it tries to apply unification rules on the problem [t1 ≡ t2]. If
    no unification rule applies then it adds [t1 = t2] in the unsolved problems
    of [p]. *)
 let add_to_unsolved : ctxt -> term -> term -> problem -> problem =
   fun ctx t1 t2 p ->
-  if Eval.eq_modulo ctx t1 t2 then p else
-  match try_unif_rules ctx t1 t2 with
-  | None     -> {p with unsolved = (ctx,t1,t2) :: p.unsolved}
+  if Eval.eq_modulo ctx t1 t2 then
+    (if !log_enabled then log_unif "equivalent terms"; p)
+  else match try_unif_rules ctx t1 t2 with
+  | None ->
+      if !log_enabled then log_unif "move to unsolved";
+      {p with unsolved = (ctx,t1,t2) :: p.unsolved}
   | Some([]) -> assert false
   (* Unification rules generate non empty list of unification constraints *)
-  | Some(cs) -> {p with to_solve = cs @ p.to_solve}
+  | Some(cs) ->
+      {p with to_solve = List.fold_left add_unif_rule_constr p.to_solve cs}
 
-(** [decompose ctx t1 ts1 t2 ts2] tries to decompose a problem of the form [h
-   ts1 = h ts2] into the problems [t1 = u1; ..; tn = un], assuming that [ts1 =
+(** [decompose ctx ts1 ts2] tries to decompose a problem of the form [h ts1 ≡
+   h ts2] into the problems [t1 ≡ u1; ..; tn ≡ un], assuming that [ts1 =
    [t1;..;tn]] and [ts2 = [u1;..;un]]. *)
-let decompose :
-      ctxt -> term -> term list -> term -> term list -> problem -> problem =
-  fun ctx t1 ts1 t2 ts2 p ->
-    if !log_enabled then log_unif "decompose";
-    let add_constr a b p = (ctx,a,b)::p in
-    let to_solve =
-      (* we use fold_right2 instead of fold_left2 to keep the order of the
-         decomposition: f a b ≡ f a' b' => a ≡ a && b ≡ b' *)
-      try List.fold_right2 add_constr ts1 ts2 p.to_solve
-      with Invalid_argument _ -> error t1 t2 in
-    {p with to_solve}
+let decompose : ctxt -> term list -> term list -> problem -> problem =
+  fun ctx ts1 ts2 p ->
+    if !log_enabled && ts1 <> [] then log_unif "decompose";
+    let add_arg_constr a b cs = add_constr (ctx,a,b) cs in
+    (* we use fold_right2 instead of fold_left2 to keep the order of the
+       decomposition: f a b ≡ f a' b' => a ≡ a && b ≡ b' *)
+    {p with to_solve = List.fold_right2 add_arg_constr ts1 ts2 p.to_solve}
 
-(** For a problem [m[vs]=s(ts)] in context [ctx], where [vs] are distinct
+(** [add_constr c p] adds [c] into [!initial] and [p]. *)
+let add_constr : constr -> problem -> problem = fun c p ->
+  {p with to_solve = c::p.to_solve}
+
+(** For a problem [m[vs] ≡ s(ts)] in context [ctx], where [vs] are distinct
    variables, [m] is a meta of type [Πy0:a0,..,Πyk-1:ak-1,b] with [k = length
    vs], [s] is an injective symbol of type [Πx0:b0,..,Πxn-1:bn-1,c] with [n =
    length ts], [imitate_inj ctx m vs us s ts] tries to instantiate [m] by
@@ -245,14 +241,14 @@ let imitate_lam_cond : term -> term list -> bool = fun h ts ->
       | Vari x -> not (Bindlib.occur x (lift h))
       | _ -> false
 
-(** For a problem of the form [Appl(m[ts],[Vari x;_])=_], where [m] is a
+(** For a problem of the form [Appl(m[ts],[Vari x;_]) ≡ _], where [m] is a
    metavariable of arity [n] and type [Πx1:a1,..,Πxn:an,t], and [x] does not
    occur in [m[ts]], instantiate [m] by [λx1:a1,..,λxn:an,λx:a,m1[x1,..,xn,x]]
    where [m1] is a new metavariable of arity [n+1] and:
 
   - either [t = Πx:a,b] and [m1] is of type [Πx1:a1,..,Πxn:an,Πx:a,b]
 
-  - or we add the problem [t = Πx:m2[x1,..,xn],m3[x1,..,xn,x]] where [m2] is a
+  - or we add the problem [t ≡ Πx:m2[x1,..,xn],m3[x1,..,xn,x]] where [m2] is a
    new metavariable of arity [n] and type [Πx1:a1,..,Πxn:an,TYPE] and [m3] is
    a new metavariable of arity [n+1] and type
    [Πx1:a1,..,Πxn:an,Πx:m2[x1,..,xn],TYPE], and do as in the previous case. *)
@@ -280,7 +276,7 @@ let imitate_lam : ctxt -> meta -> problem -> problem = fun ctx m p ->
          let b = _Meta m3 (Env.to_tbox env') in
          (* Could be optimized by extending [Env.to_tbox env]. *)
          let u = Bindlib.unbox (_Prod a (Bindlib.bind_var x b)) in
-         x,a,env',b,{p with to_solve = (Env.to_ctxt env,u,t)::p.to_solve}
+         x,a,env',b,add_constr (Env.to_ctxt env,u,t) p
     in
     let tm1 = Env.to_prod env' b in
     let m1 = Meta.fresh tm1 (n+1) in
@@ -289,111 +285,34 @@ let imitate_lam : ctxt -> meta -> problem -> problem = fun ctx m p ->
     let v = Bindlib.bind_mvar (Env.vars env) xu1 in
     Meta.set m (Bindlib.unbox v); p
 
-(** [inverses_for_prod s] returns the list of triples [(s0,s1,s2,b)] such that
-   [s] has a rule of the form [s(s0 l1 l2) ↪ Πx:s1(r1),s2(r2)] with [b=true]
-   iff [x] occurs in [r2]. *)
-type inverse = sym * sym * sym * bool
-
-let inverses_for_prod : sym -> inverse list = fun s ->
-  let f l rule =
-    match rule.lhs with
-    | [l1] ->
-        begin
-          match LibTerm.get_args l1 with
-          | Symb(s0), [_;_] ->
-              let n = Bindlib.mbinder_arity rule.rhs in
-              begin
-                match Bindlib.msubst rule.rhs (Array.make n TE_None) with
-                | Prod (Appl (Symb(s1), _), b) ->
-                    begin
-                      match Bindlib.subst b Kind with
-                      | Appl (Symb(s2), Appl(_,Kind)) -> (s0,s1,s2,true)::l
-                      | Appl (Symb(s2), _) -> (s0,s1,s2,false)::l
-                      | _ -> l
-                    end
-                | _ -> l
-              end
-          | _, _ -> l
-        end
-    | _ -> l
-  in
-  let l = List.fold_left f [] !(s.sym_rules) in
-  if !log_enabled then
-    (let f (s0,s1,s2,b) =
-       log_unif (yel "inverses_for_prod %a: %a, %a, %a, %b")
-         pp_sym s pp_sym s0 pp_sym s1 pp_sym s2 b
-     in List.iter f l);
-  l
-
-(** [inverse s v] computes [s^{-1}(v)], that is, a term [u] such that [s(u)]
-   reduces to [v], or raises [Not_invertible]. *)
-exception Not_invertible
-
-let rec inverse : sym -> term -> term = fun s v ->
-  if !log_enabled then log_unif "inverse [%a] [%a]" pp_sym s pp_term v;
-  match LibTerm.get_args (Eval.whnf [] v) with
-  | Symb(s'), [u] when s' == s -> u
-  | Prod(a,b), _ -> find_inverse_prod a b (inverses_for_prod s)
-  | _, _ -> raise Not_invertible
-
-and find_inverse_prod : term -> tbinder -> inverse list -> term = fun a b l ->
-  match l with
-  | [] -> raise Not_invertible
-  | i :: l ->
-      try inverse_prod a b i with Not_invertible -> find_inverse_prod a b l
-
-and inverse_prod : term -> tbinder -> inverse -> term =
-  fun a b (s0, s1, s2, d) ->
-  let a' = inverse s1 a in
-  let x,b = Bindlib.unbind b in
-  let b' = lift (inverse s2 b) in
-  let xb' = if d then _Abst (lift a) (Bindlib.bind_var x b') else b' in
-  add_args (Symb s0) [a'; Bindlib.unbox xb']
-
-(** [inverse_opt s ts v] returns [Some(t, s^{-1}(v))] if [ts=[t]], [s] is
-   injective and [s^{-1}(v)] exists, and [None] otherwise. *)
+(** [inverse_opt s ts v] returns [Some(t, inverse s v)] if [ts=[t]], [s] is
+   injective and [inverse s v] does not fail, and [None] otherwise. *)
 let inverse_opt : sym -> term list -> term -> (term * term) option =
   fun s ts v ->
-  if is_injective s then
+  if !log_enabled then log_unif "try inverse";
+  try
     match ts with
-    | [t] -> (try Some (t, inverse s v) with Not_invertible -> None)
-    | _ -> None
-  else None
+    | [t] when is_injective s -> Some (t, Inverse.inverse s v)
+    | _ -> raise Not_found
+  with Not_found -> if !log_enabled then log_unif "failed"; None
 
-(** [add_inverse t1 s ts1 t2] tries to replace a problem of the form [t1 = t2]
-   with [t1 = s(ts1)] by [t1 = s^{-1}(t2)] when [s] is injective. *)
+(** [add_inverse t1 s ts1 t2] tries to replace a problem of the form [t1 ≡ t2]
+   with [t1 = s(ts1)] and [ts1=[u]] by [u ≡ inverse s t2], when [s] is
+   injective. *)
 let add_inverse :
       ctxt -> term -> sym -> term list -> term -> problem -> problem =
   fun ctx t1 s ts1 t2 p ->
-  if !log_enabled then log_unif "add_inverse %a ≡ %a" pp_term t1 pp_term t2;
   match inverse_opt s ts1 t2 with
-  | Some (a, b) -> {p with to_solve = (ctx,a,b)::p.to_solve}
-  | None -> add_to_unsolved ctx t1 t2 p
+  | Some (t, u) -> add_constr (ctx,t,u) p
+  | _ -> add_to_unsolved ctx t1 t2 p
 
-(** For a problem of the form [m[ts] = Πx:_,_] with [ts] distinct bound
-   variables, [imitate_prod m ts] instantiates [m] by a fresh product. *)
+(** For a problem of the form [h1 ≡ h2] with [h1 = m[ts]], [h2 = Πx:_,_] (or
+   the opposite) and [ts] distinct bound variables, [imitate_prod m h1 h2 p]
+   instantiate [m] to a product and add the problem [h1 ≡ h2] to [p]. *)
 let imitate_prod : ctxt -> meta -> term -> term -> problem -> problem =
   fun ctx m h1 h2 p ->
   if !log_enabled then log_unif "imitate_prod %a" pp_meta m;
-  let n = m.meta_arity in
-  let (env, s) = Env.of_prod ctx n !(m.meta_type) in
-  let xs = Array.map _Vari (vars env) in
-
-  let u1 = to_prod env _Type in
-  let m1 = Meta.fresh u1 n in
-
-  let y = Bindlib.new_var mkfree "y" in
-  let env' = add y (_Meta m1 xs) None env in
-  let u2 = to_prod env' (lift s) in
-  let m2 = Meta.fresh u2 (n+1) in
-
-  let mxs = Bindlib.unbox (_Meta m xs) in
-  let a = _Meta m1 xs in
-  let b = Bindlib.bind_var y (_Meta m2 (Array.append xs [|_Vari y|])) in
-  let prod = Bindlib.unbox (_Prod a b) in
-
-  let ctx' = Env.to_ctxt env in
-  {p with to_solve = (ctx',mxs,prod)::(ctx,h1,h2)::p.to_solve}
+  Infer.set_to_prod m; add_constr (ctx,h1,h2) p
 
 (** [sym_sym_whnf ctx t1 s1 ts1 t2 s2 ts2 p] handles the case [s1(ts1) =
    s2(ts2); p] when [s1(ts1)] and [s2(ts2)] are in whnf. *)
@@ -403,34 +322,35 @@ let sym_sym_whnf :
   fun ctx t1 s1 ts1 t2 s2 ts2 p ->
   if s1 == s2 then
     match s1.sym_prop with
-    | Const
-    | Injec when List.same_length ts1 ts2 -> decompose ctx t1 ts1 t2 ts2 p
+    | (Const|Injec) ->
+        if List.same_length ts1 ts2 then decompose ctx ts1 ts2 p
+        else error t1 t2
     | _ -> add_to_unsolved ctx t1 t2 p
   else
     match s1.sym_prop, s2.sym_prop with
     | Const, Const -> error t1 t2
     | _, _ ->
         match inverse_opt s1 ts1 t2 with
-        | Some (t, u) -> {p with to_solve = (ctx,t,u)::p.to_solve}
-        | None ->
-            match inverse_opt s2 ts2 t1 with
-            | Some (t, u) -> {p with to_solve = (ctx,t,u)::p.to_solve}
-            | None -> add_to_unsolved ctx t1 t2 p
+        | Some (t, u) -> add_constr (ctx,t,u) p
+        | None -> add_inverse ctx t2 s2 ts2 t1 p
 
 (** [solve p] tries to solve the unification problem [p] and
     returns the constraints that could not be solved. *)
 let rec solve : problem -> constr list = fun p ->
-  if !log_enabled then log_unif "%a" pp_problem p;
   match p with
   | {to_solve = []; unsolved = []; _} -> []
   | {to_solve = []; unsolved = cs; recompute = true} ->
-     solve {empty_problem with to_solve = cs}
+      if !log_enabled then log_unif "recompute";
+      solve {empty_problem with to_solve = cs}
   | {to_solve = []; unsolved = cs; _} -> cs
-  | {to_solve = (((ctx,t1,t2)::to_solve) as initial); _} ->
+  | {to_solve = (ctx,t1,t2)::to_solve; _} ->
 
   (* We remove the first constraint from [p] for not looping. *)
   let p = {p with to_solve} in
 
+  (* We take the beta-whnf. *)
+  let t1 = Eval.whnf_beta t1 and t2 = Eval.whnf_beta t2 in
+  if !log_enabled then log_unif (gre "%a") pp_constr (ctx,t1,t2);
   let (h1, ts1) = LibTerm.get_args t1
   and (h2, ts2) = LibTerm.get_args t2 in
 
@@ -440,39 +360,15 @@ let rec solve : problem -> constr list = fun p ->
 
   | Prod(a1,b1), Prod(a2,b2)
   | Abst(a1,b1), Abst(a2,b2) ->
-     let (x,b1,b2) = Bindlib.unbind2 b1 b2 in
-     let ctx' = (x,a1,None) :: ctx in
-     (* [ts1] and [ts2] must be empty because of typing. *)
-     solve {p with to_solve = (ctx,a1,a2)::(ctx',b1,b2)::to_solve}
+      if !log_enabled then log_unif "decompose";
+      let (x,b1,b2) = Bindlib.unbind2 b1 b2 in
+      let ctx' = (x,a1,None)::ctx in
+      (* [ts1] and [ts2] must be empty because of typing or normalization. *)
+      solve (add_constr (ctx,a1,a2) (add_constr (ctx',b1,b2) p))
 
   | Vari x1, Vari x2 when Bindlib.eq_vars x1 x2 ->
-      solve (decompose ctx t1 ts1 t2 ts2 p)
-
-  | Symb s1, Symb s2
-       when s1 == s2 && s1.sym_prop <> Defin && List.same_length ts1 ts2 ->
-      solve (decompose ctx t1 ts1 t2 ts2 p)
-  | Symb s1, Symb s2
-       when s1 != s2 && s1.sym_prop = Const && s2.sym_prop = Const ->
-      error t1 t2
-
-  | Meta(m,ts), _ when ts1 = [] && instantiate ctx m ts t2 initial ->
-     solve {p with recompute = true}
-  | _, Meta(m,ts) when ts2 = [] && instantiate ctx m ts t1 initial ->
-     solve {p with recompute = true}
-
-  | Meta(m,ts), Prod _
-       when ts1 = [] && instantiation ctx m ts h2 <> None ->
-      solve (imitate_prod ctx m h1 h2 p)
-  | Prod _, Meta(m,ts)
-       when ts2 = [] && instantiation ctx m ts h1 <> None ->
-      solve (imitate_prod ctx m h1 h2 p)
-
-  | Meta(m,_), _ when imitate_lam_cond h1 ts1 ->
-      let p = imitate_lam ctx m p in
-      solve {p with to_solve = (ctx,t1,t2)::p.to_solve}
-  | _, Meta(m,_) when imitate_lam_cond h2 ts2 ->
-      let p = imitate_lam ctx m p in
-      solve {p with to_solve = (ctx,t1,t2)::p.to_solve}
+      if List.same_length ts1 ts2 then solve (decompose ctx ts1 ts2 p)
+      else error t1 t2
 
   | Type, (Kind|Prod _|Symb _|Vari _|Abst _)
   | Kind, (Type|Prod _|Symb _|Vari _|Abst _)
@@ -480,13 +376,37 @@ let rec solve : problem -> constr list = fun p ->
   | Vari _, (Type|Kind|Vari _|Prod _)
     -> error t1 t2
 
-  | _, _ ->
+  | Symb s1, Symb s2
+       when s1 == s2 && s1.sym_prop <> Defin && List.same_length ts1 ts2 ->
+      solve (decompose ctx ts1 ts2 p)
+  | Symb s1, Symb s2
+       when s1 != s2 && s1.sym_prop = Const && s2.sym_prop = Const ->
+      error t1 t2
+
+  | Meta(m,ts), _ when ts1 = [] && instantiate ctx m ts t2 ->
+     solve {p with recompute = true}
+  | _, Meta(m,ts) when ts2 = [] && instantiate ctx m ts t1 ->
+     solve {p with recompute = true}
+
+  | Meta(m,ts), Prod _ when ts1 = [] && instantiable ctx m ts h2 ->
+      solve (imitate_prod ctx m h1 h2 p)
+  | Prod _, Meta(m,ts) when ts2 = [] && instantiable ctx m ts h1 ->
+      solve (imitate_prod ctx m h1 h2 p)
+
+  | Meta(m,_), _ when imitate_lam_cond h1 ts1 ->
+      let p = imitate_lam ctx m p in solve (add_constr (ctx,t1,t2) p)
+  | _, Meta(m,_) when imitate_lam_cond h2 ts2 ->
+      let p = imitate_lam ctx m p in solve (add_constr (ctx,t1,t2) p)
+
+  | _ ->
 
   (* We reduce [t1] and [t2] and try again. *)
   let t1 = Eval.whnf ctx t1 and t2 = Eval.whnf ctx t2 in
   let (h1, ts1) = LibTerm.get_args t1
   and (h2, ts2) = LibTerm.get_args t2 in
-  let initial = (ctx,t1,t2)::to_solve in
+
+  if !log_enabled then log_unif "normalize";
+  if !log_enabled then log_unif (gre "%a") pp_constr (ctx,t1,t2);
 
   match h1, h2 with
   | Type, Type
@@ -494,51 +414,15 @@ let rec solve : problem -> constr list = fun p ->
 
   | Prod(a1,b1), Prod(a2,b2)
   | Abst(a1,b1), Abst(a2,b2) ->
-     let (x,b1,b2) = Bindlib.unbind2 b1 b2 in
-     let ctx' = (x,a1,None) :: ctx in
-     (* [ts1] and [ts2] must be empty because of typing. *)
-     solve {p with to_solve = (ctx,a1,a2)::(ctx',b1,b2)::to_solve}
+      if !log_enabled then log_unif "decompose";
+      let (x,b1,b2) = Bindlib.unbind2 b1 b2 in
+      let ctx' = (x,a1,None)::ctx in
+      (* [ts1] and [ts2] must be empty because of typing or normalization. *)
+      solve (add_constr (ctx,a1,a2) (add_constr (ctx',b1,b2) p))
 
   | Vari x1, Vari x2 when Bindlib.eq_vars x1 x2 ->
-      solve (decompose ctx t1 ts1 t2 ts2 p)
-
-  | Symb s1, Symb s2 -> solve (sym_sym_whnf ctx t1 s1 ts1 t2 s2 ts2 p)
-
-  | Meta(m,ts), _ when ts1 = [] && instantiate ctx m ts t2 initial ->
-     solve {p with recompute = true}
-  | _, Meta(m,ts) when ts2 = [] && instantiate ctx m ts t1 initial ->
-     solve {p with recompute = true}
-
-  | Meta(m,ts), Prod _
-       when ts1 = [] && instantiation ctx m ts h2 <> None ->
-      solve (imitate_prod ctx m h1 h2 p)
-  | Prod _, Meta (m,ts)
-       when ts2 = [] && instantiation ctx m ts h1 <> None ->
-      solve (imitate_prod ctx m h1 h2 p)
-
-  | Meta(m,_), _ when imitate_lam_cond h1 ts1 ->
-      let p = imitate_lam ctx m p in
-      solve {p with to_solve = (ctx,t1,t2)::p.to_solve}
-  | _, Meta(m,_) when imitate_lam_cond h2 ts2 ->
-      let p = imitate_lam ctx m p in
-      solve {p with to_solve = (ctx,t1,t2)::p.to_solve}
-
-  | Meta(m,ts), Symb s ->
-      if imitate_inj ctx m ts ts1 s ts2 then
-        solve {p with to_solve = (ctx,t1,t2)::to_solve}
-      else solve (add_to_unsolved ctx t1 t2 p)
-  | Symb s, Meta(m,ts) ->
-      if imitate_inj ctx m ts ts2 s ts1 then
-        solve {p with to_solve = (ctx,t1,t2)::to_solve}
-      else solve (add_to_unsolved ctx t1 t2 p)
-
-  | Meta _, _
-  | _, Meta _ -> solve (add_to_unsolved ctx t1 t2 p)
-
-  | Symb s, _ when not (is_constant s) ->
-      solve (add_inverse ctx t1 s ts1 t2 p)
-  | _, Symb s when not (is_constant s) ->
-      solve (add_inverse ctx t2 s ts2 t1 p)
+      if List.same_length ts1 ts2 then solve (decompose ctx ts1 ts2 p)
+      else error t1 t2
 
   | Type, (Kind|Prod _|Symb _|Vari _|Abst _)
   | Kind, (Type|Prod _|Symb _|Vari _|Abst _)
@@ -547,13 +431,46 @@ let rec solve : problem -> constr list = fun p ->
   | Abst _, (Type|Kind|Prod _)
     -> error t1 t2
 
-  | _, _ -> solve (add_to_unsolved ctx t1 t2 p)
+  | Symb s1, Symb s2 -> solve (sym_sym_whnf ctx t1 s1 ts1 t2 s2 ts2 p)
+
+  | Meta(m,ts), _ when ts1 = [] && instantiate ctx m ts t2 ->
+     solve {p with recompute = true}
+  | _, Meta(m,ts) when ts2 = [] && instantiate ctx m ts t1 ->
+     solve {p with recompute = true}
+
+  | Meta(m,ts), Prod _ when ts1 = [] && instantiable ctx m ts h2 ->
+      solve (imitate_prod ctx m h1 h2 p)
+  | Prod _, Meta (m,ts) when ts2 = [] && instantiable ctx m ts h1 ->
+      solve (imitate_prod ctx m h1 h2 p)
+
+  | Meta(m,_), _ when imitate_lam_cond h1 ts1 ->
+      let p = imitate_lam ctx m p in solve (add_constr (ctx,t1,t2) p)
+  | _, Meta(m,_) when imitate_lam_cond h2 ts2 ->
+      let p = imitate_lam ctx m p in solve (add_constr (ctx,t1,t2) p)
+
+  | Meta(m,ts), Symb s ->
+      if imitate_inj ctx m ts ts1 s ts2 then solve (add_constr (ctx,t1,t2) p)
+      else solve (add_to_unsolved ctx t1 t2 p)
+  | Symb s, Meta(m,ts) ->
+      if imitate_inj ctx m ts ts2 s ts1 then solve (add_constr (ctx,t1,t2) p)
+      else solve (add_to_unsolved ctx t1 t2 p)
+
+  | Meta _, _
+  | _, Meta _ -> solve (add_to_unsolved ctx t1 t2 p)
+
+  | Symb s, _ -> solve (add_inverse ctx t1 s ts1 t2 p)
+  | _, Symb s -> solve (add_inverse ctx t2 s ts2 t1 p)
+
+  | _ -> solve (add_to_unsolved ctx t1 t2 p)
 
 (** [solve p] tries to solve the unification problem [p] and
     returns the constraints that could not be solved.
     This is the entry point setting the flag type_check *)
 let solve : ?type_check:bool -> problem -> constr list =
-  fun ?(type_check=true) p -> Stdlib.(do_type_check := type_check); solve p
+  fun ?(type_check=true) p ->
+  if !log_enabled then log_hndl "solve %a" pp_problem p;
+  Stdlib.(do_type_check := type_check; initial := p.to_solve);
+  time_of (fun () -> solve p)
 
 (** [solve_noexn problem] attempts to solve [problem]. If there is
    no solution, the value [None] is returned. Otherwise [Some(cs)] is