+++ CHANGES.md
@@ -1,5 +1,49 @@
 ### Unreleased
 
+#### Add tactic `simpl <id>` for unfolding a specific symbol only (2021-03-22)
+
+and slightly improve `Ctxt.def_of`
+
+#### Bug fixes (2021-03-22)
+
+- fix type inference in the case of an application (t u) where the type of t is not a product yet (uncomment code commented in #596)
+- fix the order in which emacs prints hypotheses
+- fix opam dependencies: add constraint why3 <= 1.3.3
+
+#### Fix and improve inverse image computation (2021-03-16)
+
+- fix and improve in `inverse.ml` the computation of the inverse image of a term wrt an injective function (no unification rule is needed anymore in common examples, fix #342)
+- fix management of "initial" constraints in unification (initial is now a global variable updated whenever a new constraint is added)
+- when applying a unification rule, add constraints on types too (fix #466)
+- turn `Infer.make_prod` into `Infer.set_to_prod`
+- add pp_constrs for printing lists of constraints
+- make time printing optional
+- improve visualization of debugging data using colors:
+  . blue: top-level type inference/checking
+  . magenta: new constraint
+  . green: constraint to solve
+  . yellow: data from signature or context
+  . red: instantiations (and handled commands)
+
+#### Add tactic admit (2021-03-12)
+
+- rename command `admit` into `admitted`
+- `admitted`: admit the initial goal instead of the remaining goals (when the proof is an opaque definition)
+- move code on `admit` from `command.ml` to `tactic.ml`
+- add tactic `admit` (fix #380)
+  As a consequence, tactics can change the signature state now.
+
+#### Improvements in type inference, unification and printing (2021-03-11)
+
+- improve type inference and unification
+- add flag `"print_meta_args"`
+- print metavariables unescaped, add parentheses in domains
+- replace `(current_sign()).sign_path` by `current_path()`
+- improve logging:
+  . debug +h now prints data on type inference/checking
+  . provide time of type inference/checking and constraint solving
+  . give more feedback when instantiation fails
+
 #### Remove `set` keyword (2021-03-04)
 
 #### Various bug fixes (2021-03-02)
@@ -95,7 +139,7 @@
 
 Replace Earley by Menhir, Pratter and Sedlex
 
-**Syntax modifications:**
+Syntax modifications:
 
 - Add multi-lines comments `/*` ... `*/`.
 
@@ -129,7 +173,7 @@ Replace Earley by Menhir, Pratter and Sedlex
 - The minus sign `-` in the rewrite tactic has been replaced by the
   keyword `left`.
 
-**Code modifications:**
+Code modifications:
 
 - Parsing and handling are interleaved (except in the LSP server): the
   parser returns a stream of parsed commands. Requesting an item of